Попытался приблизить к MVC backend
