+++ backend/config.py
@@ -1,0 +1,5 @@
+DATABASE_URL = "sqlite+aiosqlite:///./test.db"
+
+origins = [
+    "*",  # заменить на адрес контейнера
+]
