+++ frontend/index.html
@@ -5,7 +5,7 @@
 </head>
 <body>
     <h1>Auction</h1>
-    <p>Product ID: <span id="product-id">1</span></p>
+    <p>Product ID: <span id="product-id"></span></p>
     <p>Current Price: $<span id="current-price"></span></p>
 <div class="input-group mb-3">
         <button id="decrement" class="btn btn-outline-secondary" type="button">-</button>
